+++ src/rocket/rocketcore/HoldingParts.java
@@ -1,5 +1,5 @@
 package rocket.rocketcore;
 
 public interface HoldingParts {
-    double calWeight();
+    float calWeight();
 }
