+++ src/rocket/rocketcore/rocketcomponent/RocketFuelTank.java
@@ -7,9 +7,9 @@ import rocket.rocketcore.rocketfragment.Fuel;
 
 public class RocketFuelTank extends RocketPart implements Controllable, HoldingParts {
     private Fuel fuel;
-    private double contain;
+    private float contain;
 
-    public RocketFuelTank(double fuelTankWeight, double contain, Fuel fuel) {
+    public RocketFuelTank(float fuelTankWeight, float contain, Fuel fuel) {
         super(fuelTankWeight);
         this.fuel = fuel;
         this.contain = contain;
@@ -24,23 +24,23 @@ public class RocketFuelTank extends RocketPart implements Controllable, HoldingP
         return fuel;
     }
 
-    public void burnFuel(double amount) throws IllegalArgumentException{
+    public void burnFuel(float amount) throws IllegalArgumentException{
         if (amount > contain) {
             throw new IllegalArgumentException("RFT: amount too big");
         }
         contain -= amount;
     }
 
-    public double getContain() {
+    public float getContain() {
         return contain;
     }
 
-    public void setContain(double contain) {
+    public void setContain(float contain) {
         this.contain = contain;
     }
 
     @Override
-    public double calWeight() {
+    public float calWeight() {
         return super.getWeight() + (fuel.getWeight() * contain);
     }
 }
