+++ src/rocket/rocketcore/rocketfragment/Explosive.java
@@ -3,18 +3,18 @@ package rocket.rocketcore.rocketfragment;
 import rocket.rocketcore.RocketPart;
 
 public class Explosive extends RocketPart {
-    private double power;
+    private float power;
 
-    public Explosive(double weight, double power) {
+    public Explosive(float weight, float power) {
         super(weight);
         this.power = power;
     }
 
-    public double getPower() {
+    public float getPower() {
         return power;
     }
 
-    public void setPower(double power) {
+    public void setPower(float power) {
         this.power = power;
     }
 }
